Seperated Meta declaration and definition into seperate files
The meta library has a LOT of templated code, which means we can't easily move function definitions to .cpp files. This has led to unavoidable circular includes, which was previously 'resolved' by placing the definition in a different file. For example, MetaManager::GetType was previously in MetaType.h, which meant you had to include both MetaManager.h and MetaType.h in order to call that function without the linker complaining about the missing definition.

 This has been resolved by splitting the declarations and definition into different files